+++ backend/canisters/group/api/src/queries/c2c_name_and_members.rs
@@ -1,0 +1,17 @@
+use candid::CandidType;
+use serde::{Deserialize, Serialize};
+use types::UserId;
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub struct Args {}
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub enum Response {
+    Success(SuccessResult),
+}
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub struct SuccessResult {
+    pub name: String,
+    pub members: Vec<UserId>,
+}