+++ backend/canisters/group_index/api/src/main.rs
@@ -5,6 +5,7 @@ fn main() {
     generate_candid_method!(group_index, recommended_groups, query);
     generate_candid_method!(group_index, search, query);
 
+    generate_candid_method!(group_index, delete_frozen_group, update);
     generate_candid_method!(group_index, freeze_group, update);
     generate_candid_method!(group_index, unfreeze_group, update);