+++ frontend/openchat-worker/src/worker.ts
@@ -1154,6 +1154,16 @@ self.addEventListener("message", (msg: MessageEvent<WorkerRequest>) => {
                     .catch(sendError(correlationId));
                 break;
 
+            case "deleteFrozenGroup":
+                agent.deleteFrozenGroup(payload.chatId)
+                    .then((response) =>
+                        sendResponse(correlationId, {
+                            response,
+                        })
+                    )
+                    .catch(sendError(correlationId));
+                break;
+
             case "suspendUser":
                 agent
                     .suspendUser(payload.userId, payload.reason)